handle ticket

+++ src/main/java/org/tibid/dto/BidOrderDto.java
@@ -6,9 +6,9 @@ import lombok.Data;
 public class BidOrderDto extends BaseDto {
 	private long id;
 
-	private int userId;
+	private long userId;
 
-	private int productId;
+	private long productId;
 
 	private String productName;
